Move l10n D-Bus bits to its own module
